book-cover-web-app/requirements_gradio.txt
gradio==5.0.2
opencv-python-headless==4.10.0.84
easyocr==1.7.1
torch==2.8.0
torchvision==0.23.0
Pillow==10.4.0
numpy==1.26.4
imutils==0.5.4
